+++ app/src/main/AndroidManifest.xml
@@ -29,7 +29,7 @@
             android:name=".ui.cozy_home.room_detail.UpdateCozyRoomDetailInfoActivity"
             android:exported="false" />
         <activity
-            android:name=".ui.cozy_home.roommate_detail.CozyHomeRoommateDetailActivity"
+            android:name=".ui.cozy_home.roommate.roommate_detail.CozyHomeRoommateDetailActivity"
             android:exported="false" />
         <activity
             android:name=".ui.university_certification.UniversityCertificationActivity"
@@ -73,7 +73,7 @@
             android:name=".ui.onboarding.OnboardingActivity"
             android:exported="true" />
         <activity
-            android:name=".ui.cozy_home.room.entering_room.JoinRoomActivity"
+            android:name=".ui.cozy_home.room.join_room.JoinRoomActivity"
             android:exported="true" />
         <activity
             android:name=".ui.cozy_home.room.making_room.MakingPrivateRoomActivity"
