+++ app/src/main/java/umc/cozymate/ui/cozy_home/room/making_room/MakingRoomDialogFragment.kt
@@ -1,4 +1,4 @@
-package umc.cozymate.ui.cozy_home.home
+package umc.cozymate.ui.cozy_home.room.making_room
 
 import android.app.AlertDialog
 import android.app.Dialog
@@ -15,8 +15,6 @@ import android.view.WindowManager
 import androidx.fragment.app.DialogFragment
 import umc.cozymate.R
 import umc.cozymate.databinding.FragmentMakingRoomDialogBinding
-import umc.cozymate.ui.cozy_home.room.making_room.MakingPrivateRoomActivity
-import umc.cozymate.ui.cozy_home.room.making_room.MakingPublicRoomActivity
 
 class MakingRoomDialogFragment : DialogFragment() {
     private var _binding: FragmentMakingRoomDialogBinding? = null
