+++ app/src/main/java/umc/cozymate/ui/cozy_home/roommate/roommate_recommend/RoommateRecommendVPAdapter.kt
@@ -1,4 +1,4 @@
-package umc.cozymate.ui.cozy_home.room.roommate_recommend
+package umc.cozymate.ui.cozy_home.roommate.roommate_recommend
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
