+++ app/src/main/java/umc/cozymate/ui/viewmodel/UniversityViewModel.kt
@@ -1,4 +1,4 @@
-package umc.cozymate.ui.university_certification
+package umc.cozymate.ui.viewmodel
 
 import android.content.Context
 import android.util.Log
