+++ app/src/main/res/layout/activity_cozy_home_roommate_detail.xml
@@ -12,7 +12,7 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:padding="20dp"
-        tools:context=".ui.cozy_home.roommate_detail.CozyHomeRoommateDetailActivity">
+        tools:context=".ui.cozy_home.roommate.roommate_detail.CozyHomeRoommateDetailActivity">
 
         <androidx.core.widget.NestedScrollView
             android:layout_width="match_parent"
