+++ app/src/main/res/layout/fragment_cozy_home_entering.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context=".ui.cozy_home.room.entering_room.CozyHomeEnteringInviteCodeActivity">
+    tools:context=".ui.cozy_home.room.join_room.CozyHomeEnteringInviteCodeActivity">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="wrap_content"
