+++ app/src/main/res/layout/fragment_making_room_dialog.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.cozy_home.home.MakingRoomDialogFragment">
+    tools:context=".ui.cozy_home.room.making_room.MakingRoomDialogFragment">
 
     <ImageView
         android:id="@+id/iv_x"
